pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "multi-ts"
version = "0.1.0"
description = "Multidimensional Conformal Temporal Series with Optimal Transport"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "POT>=0.9",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "tqdm",
    "joblib",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
